Set up CI config
This sets up the CI config to run tests in Github Actions. I've tweaked the
Dockerfile and docker-compose setup, so the zsh / fish config isn't run as
part of the CI setup, and this shaves about 3 minutes off the build time.

There's probably stuff we can do to speed it up, but I think this is OK for now.