Only cache for fifteen minutes
We're currently holding onto pages in the Redis cache for a week, which is
getting sent down the pipe to Cloudfront. This loosens up this restriction,
only caching pages for 15 minutes. Ideally we should have some way of
invalidating the Cloudfront cache when a page changes, but this is good
enough to solve the issue publishers are currently having