Add pagination to resource pages
This adds a simple pagination link to the resource pages. If JS is enabled, the
response is injected into the page, as well as sending the next page as a HTTP
header, so the next page link then requests the following page. If there are no
pages remaining, the next page link is removed from the DOM