Fix social media preview URLs
A call to `.url` on an image in production behaves differently to development, and
returns a full URL. This results in the root url being doubled up in production.

This adds a new tag to check if the url begins with `http`. If it does we just return
the url, if not, we prepend the image url with the site's root url.

This means the images work the same in prod and dev, and we guard against any
infrastructure changes that might happen to alter the behaviour in prod.