Increment `download_count` by one on download
This uses the `before_serve_document` hook to append `1` to the
`download_count` field each time a document is downloaded. Because this is a
blocking operation, we can use a thread to make this happen in the background.

However, as the tests run on an in-memory SQLite database, we need to make
sure this happens synchronously in test mode.

I've also bumped the `wagtail_factories` library, as this now includes a Document
factory we can use in the tests