Add title to iframes
The Wagtail embed block uses oEmbed to get embed HTML for given media URLs.
However, the YouTube oEmbed endpoint returns an iframe without a title attribute,
which breaks accessiblilty. I've added a method here to use BeautifulSoup to
append a title to the generated HTML.

BeautifulSoup is already a dependency of Wagtail, but I've added it to the
`pyproject.toml` in case it's removed at any point.