Format time to match NHS style guide
The [NHS Style Guide](https://service-manual.nhs.uk/content/numbers-measurements-dates-time) mandates that times are written with a decimal point
seperating the hour and minute, and "am" and "pm" in lower case, as well as
"midday" and "midnight" being specified for 12pm and 12am respecively. This is
different from the Django default, so I've added a template tag to get around this,
together with some tests.