Vendor in the NHS Promo content
We rely quite heavily on the NHS promo component. This has been removed in
version `> 4.x` of NHSUK Frontend, and the styles have appeared a little broken
for a while. As a quick fix, we're vendoring in the component our end, until we can
put in a more permenant fix.