Add home page and site wide banner config pages
This commit adds homepage and site wide banner config pages under the
settings section of the site, this allows a single homepage or
sitewide banner to be created and enabled on the site, each page
has a boolean enable flag, a dropdown select for the banner theme
and a wysiwyg editor that allows the use of bold, italic and link
in the banner text.

this branch is forked from the master branch, rather than dev, and
pulls in the admin config change for ga4id so that the new migrations
don't remove that config from the db.

Template changes to follow in separate commits, so we can cherry-pick
each element into the master branch if required.