Add http 301 redirects to uWSGI config
Currently these have been placed in the Dalmatian configuration,
because we cannot add redirects via the user interface. However,
this is problematic as CloudFront redirections happen via a
custom JavaScript function which must run in under 1ms. This means
that there will inevitably be a hard limit to how many redirects
we can add to the infrastructure.

Adding redirects to the uWSGI configuration at least means that
we are not abusing the CloudFront configuration, and we can add
more redirects via new PRs.