Fix the build
The build of the nhsx site was being broken by some incompatible changes
which had been made to poetry.

These changes enforce the use of poetry 1.2.2 which works with the
pyproject file as currently configured, and should allow the build
to complete allowing the site to be deployed again.

to test:
download this branch, and uncomment the line 127 in docker/web/Dockerfile
then run script/setup to build the site. confirm that the build completes.

this should confirm that the build as configured in Dockerfile-prod
will work as expected.