Dockerfile uses sudo to install envkey
The version of Alpine used as the base image here is
pinned and the envkey install script has not changed
in three years, so it is difficult to see how this
broke. However, without installing the sudo package
the sudo command in install.sh cannot run and the
Dockerfile cannot be built.

To ensure that the final image does not change we also
remove the sudo package after running the install script.