Redirects have exact source URLs
Previously, we checked that the source URL for a redirect
started at the beginning of the path in the URL. We
did not check that the end of the source path matched
the end of the URL from the HTTP request.

This meant that a long source path will sometimes have
matched with a much shorter one, for example the path:

    ^/a/b/c/page-name

might have been matched with:

    ^/a/b/c

and so the redirect returned would have been incorrect.

This commit changes all source paths in the redirects
to be more exact, so the example above would become:

    ^/a/b/c/page-name$

to avoid matching a shorter path encountered earlier
in the URLs file.

See:
    https://dxw.zendesk.com/agent/tickets/20957